+++ src/Interface/MimeTypes.php
@@ -48,32 +48,32 @@ interface MimeTypes
     /**
      * Get the first MIME type that matches the given file extension.
      *
-     * @param   string                 $extension  The file extension to check.
-     * @return  non-empty-string|null              The first matching MIME type or null if nothing matches.
+     * @param   string       $extension  The file extension to check.
+     * @return  string|null              The first matching MIME type or null if nothing matches.
      */
     public function getMimeType(string $extension): ?string;
 
     /**
      * Get the first file extension (without the dot) that matches the given MIME type.
      *
-     * @param  string                 $mimeType  The MIME type to check.
-     * @return non-empty-string|null             The first matching extension or null if nothing matches.
+     * @param  string       $mimeType  The MIME type to check.
+     * @return string|null             The first matching extension or null if nothing matches.
      */
     public function getExtension(string $mimeType): ?string;
 
     /**
      * Get all MIME types that match the given extension.
      *
-     * @param  string                  $extension  The file extension to check.
-     * @return list<non-empty-string>              An array of MIME types that match the given extension; can be empty.
+     * @param  string                      $extension  The file extension to check.
+     * @return array<int<0, max>, string>              An array of MIME types that match the given extension; can be empty.
      */
     public function getAllMimeTypes(string $extension): array;
 
     /**
      * Get all file extensions (without the dots) that match the given MIME type.
      *
-     * @param  string                  $mimeType  The MIME type to check.
-     * @return list<non-empty-string>             An array of file extensions that match the given MIME type; can be empty.
+     * @param  string                      $mimeType  The MIME type to check.
+     * @return array<int<0, max>, string>             An array of file extensions that match the given MIME type; can be empty.
      */
     public function getAllExtensions(string $mimeType): array;
 }